==> scripts/rstudioserver_analysis/k4me1_k9me3/3-filter_NA_cells_new_count_mat_for_LDA.R <==
# Jake Yeung
# Date of Creation: 2020-10-22
# File: ~/projects/scchic/scripts/rstudioserver_analysis/k4me1_k9me3/3-filter_NA_cells_new_count_mat_for_LDA.R
# description

library(dplyr)
library(tidyr)
library(ggplot2)
library(data.table)
library(Matrix)


# Load count mat ----------------------------------------------------------




# Load NA cells -----------------------------------------------------------




# Write new count mat -----------------------------------------------------
